+++ obj/disable_chkstk.cpp
@@ -1,0 +1,7 @@
+
+extern "C" {
+	void __chkstk(){
+		// disabling this function
+	};
+}
+